chore: Hide TasksSection and Notifications on Milestone V2 page
